Fix: CYCLE_TIME 자료형 수정

+++ src/main/java/kr/co/d_erp/dtos/Item.java
@@ -1,5 +1,7 @@
 package kr.co.d_erp.dtos;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -8,7 +10,8 @@ public class Item {
 	private String ITEM_CD,ITEM_NM, ITEM_FLAG, ITEM_SPEC, REMARK, UNIT_NM;
 	private String CAT_NM,ITEM_CATX1,ITEM_CATX2,CUST_NM;
 	private Integer PARENT_IDX, CUST_IDX, CAT_IDX, UNIT_IDX;
-	private Long CYCLE_TIME; // cycleTime 필요해서 Integer > Long으로 변경했습니다 문제 생기면 삭제해주세요. -민섭
+//	private Long CYCLE_TIME; // cycleTime 필요해서 Integer > Long으로 변경했습니다 문제 생기면 삭제해주세요. -민섭
+	private BigDecimal CYCLE_TIME; 
 	private Double ITEM_COST; // Integer > Double로 변경했습니다 문제 생기면 삭제해주세요. -민섭
 	
 }
